ARM: OMAP: Sync timer32k code with mainline
Sync timer32k code with mainline

+++ arch/arm/plat-omap/timer32k.c
@@ -202,7 +202,8 @@ static inline irqreturn_t _omap_32k_timer_interrupt(int irq, void *dev_id,
 	omap_32k_timer_ack_irq();
 	now = omap_32k_sync_timer_read();
 
-	while (now - omap_32k_last_tick >= OMAP_32K_TICKS_PER_HZ) {
+	while ((signed long)(now - omap_32k_last_tick)
+						>= OMAP_32K_TICKS_PER_HZ) {
 		omap_32k_last_tick += OMAP_32K_TICKS_PER_HZ;
 		timer_tick(regs);
 	}
